requirements.txt
chromadb
sentence-transformers
transformers
torch>=2.1.0
pypdf
Pillow
tqdm
numpy>=1.25.0
